cornea_app/requirements.txt
Flask==2.2.5
dash==2.11.1
dash-bootstrap-components==1.4.2
dash-mantine-components==0.12.1
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.18.3
google-cloud-storage==2.10.0
gunicorn==21.2.0
dash_auth==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
Werkzeug==2.2.3
Flask-Caching==2.1.0
pandas==2.0.3
gcsfs==2023.6.0
